Refactor to use Cobra
