+++ Dockerfile
@@ -7,7 +7,7 @@ WORKDIR /app
 COPY go.mod go.sum ./
 RUN go mod download
 
-COPY *.go ./
+COPY . ./
 RUN CGO_ENABLED=0 GOOS=linux go build -v -o nomad-pipeline
 
 FROM scratch
