+++ examples/happy-job.hcl
@@ -31,7 +31,7 @@ job "happy" {
 
       config {
         image = var.image
-        args  = ["-init"]
+        args  = ["agent", "init"]
 
         extra_hosts    = var.docker_extra_hosts
         auth_soft_fail = true
